+++ Sources/SwifterSwift/UIKit/UIApplicationExtensions.swift
@@ -55,17 +55,17 @@ public extension UIApplication {
 
     /// SwifterSwift: Application name (if applicable).
     var displayName: String? {
-        return Bundle.main.infoDictionary?[kCFBundleNameKey as String] as? String
+        return Bundle.main.displayName
     }
 
     /// SwifterSwift: App current build number (if applicable).
     var buildNumber: String? {
-        return Bundle.main.infoDictionary?[kCFBundleVersionKey as String] as? String
+        return Bundle.main.version
     }
 
     /// SwifterSwift: App's current version number (if applicable).
     var version: String? {
-        return Bundle.main.infoDictionary?[kCFBundleInfoDictionaryVersionKey as String] as? String
+        return Bundle.main.displayVersion
     }
 
 }